'add/changes-to-prisma-and-editor-logic-with-useState-now/also-fix-still-needed-for-entry-cards'
