+++ app/api/journal/[id]/route.ts
@@ -1,3 +1,4 @@
+import { analyze } from "@/utils/ai";
 import { getUserByClerkID } from "@/utils/auth";
 import { prisma } from "@/utils/db";
 import { NextResponse } from "next/server";
@@ -17,5 +18,18 @@ export const PATCH = async (request: Request, { params }) => {
     },
   });
 
-  return NextResponse.json({ data: updatedEntry });
+  const analysis = await analyze(updatedEntry.content);
+
+  const updated = await prisma.analysis.upsert({
+    where: {
+      entryId: updatedEntry.id,
+    },
+    create: {
+      entryId: updatedEntry.id,
+      ...analysis,
+    },
+    update: analysis,
+  });
+
+  return NextResponse.json({ data: { ...updatedEntry, analysis: updated } });
 };
